requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=10.0.0
streamlit-webrtc>=0.47.0
av>=10.0.0